backend/requirements.txt
# Flask and extensions
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0

# Database
SQLAlchemy==2.0.27

# Testing
pytest==8.0.2
pytest-cov==4.1.0
coverage==7.4.3

# Utilities
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
python-dateutil==2.9.0
gunicorn==21.2.0
eventlet==0.35.1

# Development
black==24.2.0
flake8==7.0.0
isort==5.13.2
